move DigitalOcean engine configuration validation to separate file
